analyzed cluster now also contains buffer EC2s
